+++ terminal/width.go
@@ -12,3 +12,15 @@ func detectWrapWidth(w io.Writer) int {
 	}
 	return terminalWidth(f)
 }
+
+func isTerminal(w io.Writer) bool {
+	f, ok := w.(*os.File)
+	if !ok {
+		return false
+	}
+	info, err := f.Stat()
+	if err != nil {
+		return false
+	}
+	return info.Mode()&os.ModeCharDevice != 0
+}